% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logistic}
\alias{logistic}
\title{Logistic Regression Prediction Function}
\usage{
logistic(X, beta)
}
\arguments{
\item{X}{A numeric matrix where each row is an observation, and each column is a feature.}

\item{beta}{A numeric vector of coefficients for the logistic regression model.}
}
\value{
A numeric vector of predicted probabilities.
}
\description{
This function calculates the predicted probabilities using a logistic regression model.
}
