% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_PimaIndiansDiabetes.R
\docType{data}
\name{PimaIndiansDiabetes}
\alias{PimaIndiansDiabetes}
\title{PimaIndiansDiabetes Data Set}
\format{
A data frame with 768 rows and 9 variables:
\describe{
\item{pregnant}{Number of times pregnant}
\item{glucose}{Plasma glucose concentration}
\item{pressure}{Diastolic blood pressure}
\item{triceps}{Triceps skin fold thickness}
\item{insulin}{2-Hour serum insulin}
\item{mass}{Body mass index}
\item{pedigree}{Diabetes pedigree function}
\item{age}{Age in years}
\item{diabetes}{Class variable (0 or 1)}
}
}
\source{
{Smith et al. (1988). Using the ADAP learning algorithm to forecast the onset of diabetes mellitus.}
}
\usage{
data(PimaIndiansDiabetes)
}
\description{
A dataset containing medical records for Pima Indian women, used to predict diabetes.
}
\keyword{datasets}
