% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J_log_likelihood.R
\name{J_log_likelihood}
\alias{J_log_likelihood}
\title{Log-Likelihood Function for Logistic Regression}
\usage{
J_log_likelihood(X, y, beta)
}
\arguments{
\item{X}{A numeric matrix where each row represents an observation and each column represents a feature.}

\item{y}{A numeric vector of binary outcomes (0 or 1) for each observation.}

\item{beta}{A numeric vector of regression coefficients.}
}
\value{
The negative log-likelihood value.
}
\description{
This function computes the negative log-likelihood for a logistic regression model.
}
