% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J_logistic_regression.R
\name{J_logistic}
\alias{J_logistic}
\title{Logistic Regression Prediction Function}
\usage{
J_logistic(X, beta)
}
\arguments{
\item{X}{A numeric matrix where each row is an observation, and each column is a feature.}

\item{beta}{A numeric vector of coefficients for the logistic regression model.}
}
\value{
A numeric vector of predicted probabilities.
}
\description{
This function calculates the predicted probabilities using a logistic regression model.
}
\examples{
X <- matrix(c(1, 2, 3, 4, 5, 6), nrow = 3, byrow = TRUE)
beta <- c(0.5, -0.2)
J_logistic(X, beta)
#The result for this example: 0.5249792 0.6681878 0.7858350
}
